Move everything to ModelStringUtil
