+++ src/main/java/net/sf/jabref/cli/AuxCommandLine.java
@@ -2,15 +2,15 @@ package net.sf.jabref.cli;
 
 import net.sf.jabref.logic.auxparser.AuxParser;
 import net.sf.jabref.logic.auxparser.AuxParserResult;
-import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.model.database.BibDatabase;
+import net.sf.jabref.model.util.ModelStringUtil;
 
 public class AuxCommandLine {
     private final String auxFile;
     private final BibDatabase database;
 
     public AuxCommandLine(String auxFile, BibDatabase database) {
-        this.auxFile = StringUtil.getCorrectFileName(auxFile, "aux");
+        this.auxFile = ModelStringUtil.getCorrectFileName(auxFile, "aux");
         this.database = database;
     }
