+++ src/main/java/net/sf/jabref/gui/groups/GroupTreeCellRenderer.java
@@ -13,8 +13,8 @@ import javax.swing.JTree;
 import javax.swing.border.Border;
 import javax.swing.tree.DefaultTreeCellRenderer;
 
-import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.model.groups.GroupTreeNode;
+import net.sf.jabref.model.util.ModelStringUtil;
 
 /**
  * Renders a GroupTreeNode using its group's getName() method, rather that its toString() method.
@@ -71,7 +71,7 @@ public class GroupTreeCellRenderer extends DefaultTreeCellRenderer {
         if (viewModel.printInItalics()) {
             sb.append("<i>");
         }
-        sb.append(StringUtil.quoteForHTML(viewModel.getText()));
+        sb.append(ModelStringUtil.quoteForHTML(viewModel.getText()));
         if (viewModel.printInItalics()) {
             sb.append("</i>");
         }
