+++ src/main/java/net/sf/jabref/gui/importer/actions/OpenDatabaseAction.java
@@ -30,13 +30,13 @@ import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.util.FileExtensions;
 import net.sf.jabref.logic.util.io.AutoSaveUtil;
 import net.sf.jabref.logic.util.io.FileBasedLock;
-import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.migrations.FileLinksUpgradeWarning;
 import net.sf.jabref.model.Defaults;
 import net.sf.jabref.model.database.BibDatabase;
 import net.sf.jabref.model.database.BibDatabaseContext;
 import net.sf.jabref.model.database.BibDatabaseMode;
 import net.sf.jabref.model.metadata.MetaData;
+import net.sf.jabref.model.util.ModelStringUtil;
 import net.sf.jabref.preferences.JabRefPreferences;
 
 import org.apache.commons.logging.Log;
@@ -85,7 +85,7 @@ public class OpenDatabaseAction extends MnemonicAwareAction {
             filesToOpen.addAll(chosenStrings.stream().map(File::new).collect(Collectors.toList()));
         } else {
             LOGGER.info(Action.NAME + " " + e.getActionCommand());
-            filesToOpen.add(new File(StringUtil.getCorrectFileName(e.getActionCommand(), "bib")));
+            filesToOpen.add(new File(ModelStringUtil.getCorrectFileName(e.getActionCommand(), "bib")));
         }
 
         openFiles(filesToOpen, true);
