+++ src/main/java/net/sf/jabref/gui/undo/UndoableInsertString.java
@@ -2,10 +2,10 @@ package net.sf.jabref.gui.undo;
 
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.logic.l10n.Localization;
-import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.model.database.BibDatabase;
 import net.sf.jabref.model.database.KeyCollisionException;
 import net.sf.jabref.model.entry.BibtexString;
+import net.sf.jabref.model.util.ModelStringUtil;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -28,7 +28,7 @@ public class UndoableInsertString extends AbstractUndoableJabRefEdit {
 
     @Override
     public String getPresentationName() {
-        return Localization.lang("insert string %0", StringUtil.boldHTML(string.toString()));
+        return Localization.lang("insert string %0", ModelStringUtil.boldHTML(string.toString()));
     }
 
     @Override
