+++ src/main/java/net/sf/jabref/gui/undo/UndoableStringChange.java
@@ -2,8 +2,8 @@ package net.sf.jabref.gui.undo;
 
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.logic.l10n.Localization;
-import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.model.entry.BibtexString;
+import net.sf.jabref.model.util.ModelStringUtil;
 
 public class UndoableStringChange extends AbstractUndoableJabRefEdit {
 
@@ -26,9 +26,9 @@ public class UndoableStringChange extends AbstractUndoableJabRefEdit {
 
     @Override
     public String getPresentationName() {
-        return (nameChange ? Localization.lang("change string name %0 to %1", StringUtil.boldHTML(oldValue),
-                StringUtil.boldHTML(newValue)) : Localization.lang("change string content %0 to %1",
-                        StringUtil.boldHTML(oldValue), StringUtil.boldHTML(newValue)));
+        return (nameChange ? Localization.lang("change string name %0 to %1", ModelStringUtil.boldHTML(oldValue),
+                ModelStringUtil.boldHTML(newValue)) : Localization.lang("change string content %0 to %1",
+                        ModelStringUtil.boldHTML(oldValue), ModelStringUtil.boldHTML(newValue)));
     }
 
     @Override
