+++ src/main/java/net/sf/jabref/logic/bibtex/FieldContentParser.java
@@ -5,8 +5,9 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.regex.Pattern;
 
-import net.sf.jabref.logic.util.strings.StringUtil;
+import net.sf.jabref.logic.util.OS;
 import net.sf.jabref.model.entry.FieldName;
+import net.sf.jabref.model.util.ModelStringUtil;
 
 /**
  * This class provides the reformatting needed when reading BibTeX fields formatted
@@ -43,7 +44,7 @@ public class FieldContentParser {
 
         if (multiLineFields.contains(bibtexField)) {
             // Unify line breaks
-            return StringUtil.unifyLineBreaksToConfiguredLineBreaks(fieldContent);
+            return ModelStringUtil.unifyLineBreaksToConfiguredLineBreaks(fieldContent, OS.NEWLINE);
         }
 
         return WHITESPACE.matcher(fieldContent).replaceAll(" ");
