+++ src/main/java/net/sf/jabref/logic/bibtex/LatexFieldFormatter.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import net.sf.jabref.logic.util.OS;
-import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.model.entry.InternalBibtexFields;
+import net.sf.jabref.model.util.ModelStringUtil;
 
 /**
  * Currently the only implementation of net.sf.jabref.exporter.FieldFormatter
@@ -250,7 +250,7 @@ public class LatexFieldFormatter {
     }
 
     private void putIn(String s) {
-        stringBuilder.append(StringUtil.wrap(s, prefs.getLineLength()));
+        stringBuilder.append(ModelStringUtil.wrap(s, prefs.getLineLength(), OS.NEWLINE));
     }
 
     private static void checkBraces(String text) throws IllegalArgumentException {
