+++ src/main/java/net/sf/jabref/logic/groups/GroupDescriptions.java
@@ -1,18 +1,18 @@
 package net.sf.jabref.logic.groups;
 
 import net.sf.jabref.logic.l10n.Localization;
-import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.model.groups.ExplicitGroup;
 import net.sf.jabref.model.groups.KeywordGroup;
 import net.sf.jabref.model.groups.SearchGroup;
+import net.sf.jabref.model.util.ModelStringUtil;
 
 public class GroupDescriptions {
 
     public static String getDescriptionForPreview(String field, String expr, boolean caseSensitive, boolean regExp) {
         String header = regExp ? Localization.lang("This group contains entries whose <b>%0</b> field contains the regular expression <b>%1</b>",
-                field, StringUtil.quoteForHTML(expr))
+                field, ModelStringUtil.quoteForHTML(expr))
                 : Localization.lang("This group contains entries whose <b>%0</b> field contains the keyword <b>%1</b>",
-                field, StringUtil.quoteForHTML(expr));
+                field, ModelStringUtil.quoteForHTML(expr));
         String caseSensitiveText = caseSensitive ? Localization.lang("case sensitive") :
             Localization.lang("case insensitive");
         String footer = regExp ?
@@ -27,7 +27,7 @@ public class GroupDescriptions {
                         + "then using the context menu. "
                         + "This process removes the term <b>%1</b> from "
                         + "each entry's <b>%0</b> field.",
-                field, StringUtil.quoteForHTML(expr));
+                field, ModelStringUtil.quoteForHTML(expr));
         return String.format("%s (%s). %s", header, caseSensitiveText, footer);
     }
 
@@ -35,9 +35,9 @@ public class GroupDescriptions {
         StringBuilder sb = new StringBuilder();
         sb.append("<b>");
         if (showDynamic) {
-            sb.append("<i>").append(StringUtil.quoteForHTML(keywordGroup.getName())).append("</i>");
+            sb.append("<i>").append(ModelStringUtil.quoteForHTML(keywordGroup.getName())).append("</i>");
         } else {
-            sb.append(StringUtil.quoteForHTML(keywordGroup.getName()));
+            sb.append(ModelStringUtil.quoteForHTML(keywordGroup.getName()));
         }
         sb.append("</b> - ");
         sb.append(Localization.lang("dynamic group"));
@@ -46,7 +46,7 @@ public class GroupDescriptions {
         sb.append("</b> ");
         sb.append(Localization.lang("contains"));
         sb.append(" <b>");
-        sb.append(StringUtil.quoteForHTML(keywordGroup.getSearchExpression()));
+        sb.append(ModelStringUtil.quoteForHTML(keywordGroup.getSearchExpression()));
         sb.append("</b>)");
         switch (keywordGroup.getHierarchicalContext()) {
         case INCLUDING:
@@ -95,15 +95,15 @@ public class GroupDescriptions {
         StringBuilder sb = new StringBuilder();
         sb.append("<b>");
         if (showDynamic) {
-            sb.append("<i>").append(StringUtil.quoteForHTML(searchGroup.getName())).append("</i>");
+            sb.append("<i>").append(ModelStringUtil.quoteForHTML(searchGroup.getName())).append("</i>");
         } else {
-            sb.append(StringUtil.quoteForHTML(searchGroup.getName()));
+            sb.append(ModelStringUtil.quoteForHTML(searchGroup.getName()));
         }
         sb.append("</b> - ");
         sb.append(Localization.lang("dynamic group"));
         sb.append(" (");
         sb.append(Localization.lang("search expression"));
-        sb.append(" <b>").append(StringUtil.quoteForHTML(searchGroup.getSearchExpression())).append("</b>)");
+        sb.append(" <b>").append(ModelStringUtil.quoteForHTML(searchGroup.getSearchExpression())).append("</b>)");
         switch (searchGroup.getHierarchicalContext()) {
         case INCLUDING:
             sb.append(", ").append(Localization.lang("includes subgroups"));
